resume/requirements.txt
fastapi
uvicorn
openai
jinja2
pydantic
groq
python-multipart
google-cloud-storage
